refactor: Introduced an utils module
